Simplify `It.IsAny<T>`, `It.IsNotNull<T>` matchers
by documenting that `IsAssignableFrom` generally yields the same
results as the `is` operator when run on an `object`-typed variable.

Note that this is removing the `IsComObject` check as well. This check
was originally added because it was found that `IsAssignableFrom` does
not always work properly for COM interfaces whereas the `is` works.
Since we're generally going back on `is`, the COM check becomes redun-
dant.