Generic args must match exactly for inner mock setups
While this fixes the regression documented in the previous commit, it
might be an insufficent solution in general. Perhaps it would be more
correct to take co-/contra-variance into account and select a compari-
son mode per generic argument depending on whether it appears in a
input and/or output position?