Remove obsolete hack which causes this problem
One side effect of `FluentMockVisitor` (recently removed) was that it
caused extra invocations to be recorded during mock setup. We then had
to delete that record to give off the impression that new mocks are
truly untouched.

Unfortunately, this also caused invocations to be cleared that should
have stayed in the record.

We no longer use `FluentMockVisitor`, so no extra invocations are re-
corded in the first place.