Replace `Guard.CanBeAssigned`
Its only extant use was to check whether a type implements an inter-
face, so turn it into a more specific method `ImplementsInterface`.
We can then further simplify another `Guard` method.

(As an aside, this resolves an old TODO: The `Guard` call in the ctor
of `MockQueryable<T>` cannot be turned into a `Debug.Assert`, as it
prevents external code (such as a rogue LINQ query operator) from con-
structing an invalid `MockQueryable<T>` via `CreateQuery` when it
should have `Execute`d instead.)