Expose `Verify[All]` methods
Despite the same name, they differ from the `Verify[All]` methods of
`Mock` mainly in one aspect: The setup on which the method is invoked
will always be verified, regardless of whether or not it is flagged
verifiable.

That is, `setup.Verify()` will verify `setup` even if it has not been
marked `.Verifiable()`. It wouldn't make sense for `setup.Verify()` to
be a no-op; otherwise, why would one make the call in the first place?

These new method signatures have the added benefit of allowing us to
later introduce (both on `ISetup` and `Mock`) an additional overload:

    void Verify(bool recursive, Func<ISetup, bool> predicate)

which will allow custom setup selection without conflicting with any
existing `Verify` methods.