Mark matched invocations as verified
The exact location where this `foreach` loop gets added is important:

The simplest solution with the fewest lines of code would be to mark
matching invocations inside `Setup.TryVerify`, and to remove the corr-
esponding loop from `Mock.TryVerify`.

However, since `Setup.TryVerify` is called by `Mock.TryVerify` for each
setup, that would mean iterating over invocations once per setup. It
is much more efficient to iterate once per mock in `Setup.TryVerify`.

This is why we are adding the new `foreach` loop in `Setup.Verify`, so
it only executes when we enter verification at this setup (and so "by-
passed" the batch logic in `Mock`).