Missing `invocation.Return()` for `void` methods
`Invocation` expects to be "terminated" by a call to any of its
`Return` methods, yet that currently never happens for `void` methods
that have a setup but no further behavior configured.

Add the missing call.

The necessary guard condition is more complex than it needs to be,
let's clean that up in the following commits.