Prevent argument count mismatch
The expression splitting logic is a little hacky when it comes to
assignments to property or indexer setters: first, a split operation
is performed as if we were looking at a property or indexer *getter*;
the `InvocationShape` receives the correct method, but is one argument
short. Once that first `InvocationShape` has been built, we can create
a second, patched and fully correct version that includes the assign-
ment (RHS) argument.

This commit relaxes validation rules for the first `InvocationShape`
(i.e. the one with the potential argument count mismatch).