Unwrap double (and triple, etc.) conversions
The existing code in `MatcherFactory` unwraps only one conversion node
to get at matchers etc. This is not enough in the scenario documented
in the previous commit's failing test, which causes two conversion nodes
to be present.

This bug fix assumes that any conversion nodes present are only there
to satisfy the compiler's static type checks, but that they are irrele-
vant (and can thus be removed) for any analysis that needs to be done
to identify a suitable matcher.

(It might be safer to not simply remove any conversion node blindly,
but instead target this very specific VB.NET pattern and deal with it
as a special case.)