Make `DefaultValue.Mock` inherit property stubbing
The XML doc comment for `Mock<T>.SetupAllProperties()` states that:

> If the mock's `DefaultValue` is set to `DefaultValue.Mock`, the
> mocked default values will also get all properties setup recursively.

But this does not actually happen! This commit changes the behavior of
`DefaultValue.Mock` such that it does what it says on the tin.

In theory, this could be a breaking change; let's see how many users
have relied on behavior that contradicts the documentation.

References:
https://github.com/moq/moq4/blob/947c6a0aa1bbf265b5555ad0a84beb5bfc53b4e2/src/Moq/Mock.Generic.cs#L671-L674