New helper method `SubstituteTypeMatchers`
When testing an argument's type against a "template type" that may in-
volve type matchers, we can no longer do a simple `IsAssignableFrom`;
we'd have to create our own type matcher-aware version of it that com-
pares each constituent part of both sides' types (visited in lockstep).

Implementing our own `IsAssignableFrom` is going to be quite difficult
because we also have to account for variance. Looking at both ECMA-335
and .NET Core's source code, that looks rather complex... let's not go
there.

What we'll do instead is much easier: we still visit both sides' types'
constituent parts, and whenever we encounter a type matcher in the
template type, we test it against the corresponding (sub-) type from
the argument. If it's a match, we substitute the argument type in the
template type. If all type matchers match, we end up with a rewritten
template type that no longer uses any matchers; we can then use `Is-
AssignableFrom` with _that_ template type and the argument type to let
the framework deal with variance.