Don't depend on SolutionDir to allow Pack from project
The SolutionDir property is only available when building from the
solution. For packing, it's actually not necessary and we could
also easily allow packing from the Moq.csproj project itself.

This change, in addition to enabling that, also makes the icon
path more flexible since it could be moved down to the src folder
(say) without requiring a change to the .csproj.