Embed symbols in assembly, drop .snupkg
Since we're generating portable symbols, these are WAY smaller than
full PDBs used to be. Their size contribution to the package are
negligible and for convenience, it's more common nowadays to just
embed the symbols always in the assembly itself. In particular for
a library that is not deployed to production apps, this makes
perfect sense.

So stop generating the (effectively legacy) .snupkg.