Convert `FindMatchFor` to `FindLast`
By accepting a `Setup` predicate instead of an `Invocation`, the method
becomes more generally useful.

Adding `Last` to the name clarifies that the collection will be searched
in reverse order. This needs to be clarified because some methods
process the collection in forward order, while others do the opposite.

The name change also makes `SetupCollection` more similar to regular
.NET collections (see e.g. `List<>.FindLast`).